+++ packages/viz/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [5.13.13](https://github.com/lumada-design/hv-uikit-react/compare/@hitachivantara/uikit-react-viz@5.13.12...@hitachivantara/uikit-react-viz@5.13.13) (2024-08-02)
+
+**Note:** Version bump only for package @hitachivantara/uikit-react-viz
+
+
+
+
+
 ## [5.13.12](https://github.com/lumada-design/hv-uikit-react/compare/@hitachivantara/uikit-react-viz@5.13.11...@hitachivantara/uikit-react-viz@5.13.12) (2024-08-01)
 
 **Note:** Version bump only for package @hitachivantara/uikit-react-viz